+++ example/lib/main.dart
@@ -4,6 +4,7 @@ import 'package:sahha_flutter_example/Views/AuthenticationView.dart';
 import 'package:sahha_flutter_example/Views/BiomarkersView.dart';
 import 'package:sahha_flutter_example/Views/HomeView.dart';
 import 'package:sahha_flutter_example/Views/ProfileView.dart';
+import 'package:sahha_flutter_example/Views/SamplesView.dart';
 import 'package:sahha_flutter_example/Views/ScoresView.dart';
 import 'package:sahha_flutter_example/Views/SensorPermissionView.dart';
 import 'package:sahha_flutter_example/Views/StatsView.dart';
@@ -62,6 +63,7 @@ class AppState extends State<App> {
         '/scores': (BuildContext context) => const ScoresView(),
         '/biomarkers': (BuildContext context) => const BiomarkersView(),
         '/stats': (BuildContext context) => const StatsView(),
+        '/samples': (BuildContext context) => const SamplesView(),
         '/web': (BuildContext context) => const WebView(),
       },
     );